Move loadGLTF to utils
