+++ src/game.js
@@ -1,10 +1,10 @@
 import * as THREE from "three";
-import { GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader";
 import { GLTFExporter } from "three/examples/jsm/exporters/GLTFExporter";
 import logger from "./logger";
 import constants from "./constants";
 import assets from "./assets";
 import { exportAvatar } from "./export";
+import { loadGLTF } from "./utils";
 
 const avatarParts = Object.keys(assets);
 
@@ -37,32 +37,6 @@ document.addEventListener(constants.avatarConfigChanged, (e) => {
 document.addEventListener(constants.exportAvatar, () => {
   state.shouldExportAvatar = true;
 });
-
-const loadGLTF = (function () {
-  const loader = new GLTFLoader();
-  return function loadGLTF(url) {
-    return new Promise(function (resolve, reject) {
-      loader.load(
-        url,
-        function (gltf) {
-          resolve(gltf);
-          // gltf.animations; // Array<THREE.AnimationClip>
-          // gltf.scene; // THREE.Group
-          // gltf.scenes; // Array<THREE.Group>
-          // gltf.cameras; // Array<THREE.Camera>
-          // gltf.asset; // Object
-        },
-        function (xhr) {
-          logger.log((xhr.loaded / xhr.total) * 100 + "% loaded");
-        },
-        function (error) {
-          logger.log("An error happened");
-          reject(error);
-        }
-      );
-    });
-  };
-})();
 
 function init() {
   THREE.Cache.enabled = true;
